Restored the builder to its former glory + more... While DSSS integration might be nice at some future point, it'd be better to just get a simple builder done and worry about DSSS, etc. later.
Reduced references kept during the build() method of the builder. This allows the GC to collect things (like the dependancy collector cache) if memory gets tight during one of the later steps.

Removed the redundant interface IExecutionMonitor (IResponseInterpreter does the same thing... not sure which is a better name, but the code is already using IResponseInterpreter, so... also, ExecutionMonitor is already a class in the .debuild package being used for process management).

Added a simple thread-pooling mechanism to the ExecutionMonitor that uses the system's # of processors. Since DMD is single-core this will allow the compiler to take advantage of multi-core if multiple compile sessions need to be run. If it's always one huge compile orgy (a la rebuild unless -oneatatime is passed) this isn't a big issue, but if compile groups are smaller, using each CPU core for a DMD/GDC/LLVMDC process might speed up compile times.

Began implementing UI help, but I can't figure out why the context-sensitive stuff isn't working (it gives an error message if you F1 at the compiler tab), so it doesn't do anything yet. Future directions here: context-sensitive help for everything meaningful in the builder UI, TOC/traditional help for the builder and a cheat-sheet for setting up a build configuration.

git-svn-id: http://svn.dsource.org/projects/descent/trunk@1247 6538119e-2221-0410-ba12-9a5a7e5062c7